feat: Renamed `design.py` to `jobs.py` for consistency
