+++ examples/backbone_design/designs/jobs.py
@@ -1,7 +1,7 @@
 """Import designs so they are discoverable by `load_jobs`."""
 
-from .initial_data.design import InitialDesign
-from .core_site.design import CoreSiteDesign
+from .initial_data.jobs import InitialDesign
+from .core_site.jobs import CoreSiteDesign
 
 __all__ = (
     "InitialDesign",
