+++ examples/example_design/designs/jobs.py
@@ -1,5 +1,5 @@
 """Import designs so they are discoverable by `load_jobs`."""
 
-from .basic.design import BasicDesign
+from .basic.jobs import BasicDesign
 
 __all__ = ("BasicDesign",)
